+++ pages/robot/index.php
@@ -30,6 +30,14 @@ $tabs = [
 //        'name' => 'Software',
 //        'icon' => 'code-fork'
 //    ],
+    'components' => [
+        'name' => 'Components',
+        'icon' => 'puzzle-piece'
+    ],
+    'calibrations' => [
+        'name' => 'Calibrations',
+        'icon' => 'dot-circle-o'
+    ],
     'settings' => [
         'name' => 'Settings',
         'icon' => 'sliders'